linked_list_r4l
Struct RawList

pub struct RawList<G: GetLinks> { /* private fields */ }

A linked list.

Invariants

The links of objects added to a list are owned by the list.

Implementations

impl<G: GetLinks> RawList<G>

pub const fn new() -> Self

Constructs a new empty RawList.

pub fn iter(&self) -> Iterator<'_, G>

Returns an iterator for the list starting at the first entry.

pub const fn is_empty(&self) -> bool

Returns whether the RawList is empty.

pub unsafe fn insert_after( &mut self, existing: &G::EntryType, new: &G::EntryType, ) -> bool

Inserts the given object after existing.

Safety

Callers must ensure that existing points to a valid entry that is on the list.

pub unsafe fn push_back(&mut self, new: &G::EntryType) -> bool

Adds the given object to the end (back) of the list.

Rawlist will save the reference as node ptr. The caller must ensure the validity of the reference while it is on the linked list.

pub unsafe fn push_front(&mut self, new: &G::EntryType) -> bool

Adds the given object to the first (front) of the list.

Rawlist will save the reference as node ptr. The caller must ensure the validity of the reference while it is on the linked list.

pub unsafe fn remove(&mut self, data: &G::EntryType) -> bool

Removes the given entry.

Safety

Callers must ensure that data is either on this list or in no list. It being on another list leads to memory unsafety.

pub fn pop_front(&mut self) -> Option<NonNull<G::EntryType>>

Get and Remove the first element of the list.

pub fn cursor_front_mut(&mut self) -> CursorMut<'_, G>

Returns a mut cursor starting on the first element of the list.