linked_list_r4l
Trait GetLinks

pub trait GetLinks {
    type EntryType: ?Sized;

    // Required method
    fn get_links(data: &Self::EntryType) -> &Links<Self::EntryType>;
}

A descriptor of list elements.

It describes the type of list elements and provides a function to determine how to get the links to be used on a list.

A type that may be in multiple lists simultaneously needs to implement one of these for each simultaneous list.

Required Associated Types

type EntryType: ?Sized

The type of the entries in the list.

Required Methods

fn get_links(data: &Self::EntryType) -> &Links<Self::EntryType>

Returns the links to be used when linking an entry within a list.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.